flatten the files
