+++ apps/components-demo/src/pages/form/index.tsx
@@ -1,6 +1,5 @@
-import BaseLayout from '@source-cooperative/components/components/BaseLayout.js'
-import Form, { FormResult } from '@source-cooperative/components/components/Form.js'
-import { FieldState, FieldStateValue, FieldType, FormResultState } from '@source-cooperative/components/lib/enums.js'
+import BaseLayout from '@source-cooperative/components/BaseLayout.js'
+import Form, { FieldState, FieldStateValue, FieldType, FormResult, FormResultState } from '@source-cooperative/components/Form.js'
 import { useState } from 'react'
 import { Heading } from 'theme-ui'
 import { sideNavLinks } from '../../utils/constants'
