+++ apps/components-demo/src/pages/viewers/loader/index.tsx
@@ -1,6 +1,6 @@
-import BaseLayout from '@source-cooperative/components/components/BaseLayout.js'
-import ViewerLoader from '@source-cooperative/components/components/ViewerLoader.js'
-import { ViewerId } from '@source-cooperative/components/components/viewers/index.js'
+import BaseLayout from '@source-cooperative/components/BaseLayout.js'
+import ViewerLoader from '@source-cooperative/components/ViewerLoader.js'
+import { ViewerId } from '@source-cooperative/components/viewers.js'
 import { useState } from 'react'
 import { Divider, Heading, Paragraph, Select } from 'theme-ui'
 import { sideNavLinks } from '../../../utils/constants'
