+++ apps/components-demo/src/pages/viewers/map/index.tsx
@@ -1,10 +1,11 @@
-import BaseLayout from '@source-cooperative/components/components/BaseLayout.js'
-import { map } from '@source-cooperative/components/components/viewers/map/index.js'
+import BaseLayout from '@source-cooperative/components/BaseLayout.js'
+import metadata from '@source-cooperative/components/map/metadata.js'
+import MapViewer from '@source-cooperative/components/map/viewer.js'
 import { Heading, Paragraph } from 'theme-ui'
 import { sideNavLinks } from '../../../utils/constants'
 
 export default function Map() {
-  const { title, description, viewer: MapViewer } = map
+  const { title, description } = metadata
   return (
     <BaseLayout sideNavLinks={sideNavLinks}>
       <Heading as="h1">{title}</Heading>
