+++ apps/components-demo/src/pages/viewers/markdown/index.tsx
@@ -1,11 +1,12 @@
-import BaseLayout from '@source-cooperative/components/components/BaseLayout.js'
-import { markdown } from '@source-cooperative/components/components/viewers/markdown/index.js'
+import BaseLayout from '@source-cooperative/components/BaseLayout.js'
+import metadata from '@source-cooperative/components/markdown/metadata.js'
+import MarkdownViewer from '@source-cooperative/components/markdown/viewer.js'
 import { Heading, Paragraph } from 'theme-ui'
 import { sideNavLinks } from '../../../utils/constants'
 // import { markdown } from './markdown'
 
 export default function Map() {
-  const { title, description, viewer: MarkdownViewer } = markdown
+  const { title, description } = metadata
   return (
     <BaseLayout sideNavLinks={sideNavLinks}>
       <Heading as="h1">{title}</Heading>
