+++ apps/components-demo/src/pages/viewers/table/index.tsx
@@ -1,10 +1,11 @@
-import BaseLayout from '@source-cooperative/components/components/BaseLayout.js'
-import { table } from '@source-cooperative/components/components/viewers/table/index.js'
+import BaseLayout from '@source-cooperative/components/BaseLayout.js'
+import metadata from '@source-cooperative/components/table/metadata.js'
+import TableViewer from '@source-cooperative/components/table/viewer.js'
 import { Heading, Paragraph } from 'theme-ui'
 import { sideNavLinks } from '../../../utils/constants'
 
 export default function Text() {
-  const { title, description, viewer: TableViewer } = table
+  const { title, description } = metadata
   return (
     <BaseLayout sideNavLinks={sideNavLinks}>
       <Heading as="h1">{title}</Heading>
