+++ apps/viewers/src/pages/_app.tsx
@@ -1,5 +1,5 @@
-import { SourceProvider } from '@source-cooperative/components/lib/provider.js'
-import { theme } from '@source-cooperative/components/theme/index.js'
+import { SourceProvider } from '@source-cooperative/components/SourceThemeProvider.js'
+import { theme } from '@source-cooperative/components/theme.js'
 import type { AppProps } from 'next/app'
 import { SkeletonTheme } from 'react-loading-skeleton'
 import '../styles/globals.css' // TODO(SL): distribute in components package
