+++ apps/viewers/src/pages/about/index.tsx
@@ -1,4 +1,4 @@
-import Link from '@source-cooperative/components/components/Link.js'
+import Link from '@source-cooperative/components/Link.js'
 import { Heading, Paragraph } from 'theme-ui'
 import Layout from '../../components/Layout'
