+++ apps/viewers/src/pages/index.tsx
@@ -1,7 +1,6 @@
-import Form, { FormFieldState, FormResult } from '@source-cooperative/components/components/Form.js'
-import ViewerLoader from '@source-cooperative/components/components/ViewerLoader.js'
-import { ViewerId, viewerIds } from '@source-cooperative/components/components/viewers/index.js'
-import { FieldState, FieldType, FormResultState } from '@source-cooperative/components/lib/enums.js'
+import Form, { FieldState, FieldType, FormFieldState, FormResult, FormResultState } from '@source-cooperative/components/Form.js'
+import ViewerLoader from '@source-cooperative/components/ViewerLoader.js'
+import { ViewerId, viewerIds } from '@source-cooperative/components/viewers.js'
 import { useRouter } from 'next/router'
 import { useCallback, useEffect, useState } from 'react'
 import { Box, Divider } from 'theme-ui'
