+++ packages/components/src/Form.tsx
@@ -12,10 +12,32 @@ import {
   Text,
   Textarea,
 } from 'theme-ui'
-import type { FieldStateValue, FieldTypeValue, FormResultStateValue } from '../lib/enums'
-import { FieldState, FieldType, FormResultState } from '../lib/enums'
-import { SxProp } from '../lib/sx'
 import Button from './Button'
+import { SxProp } from './sx'
+
+export const FieldState = {
+  VALID: 'VALID',
+  VALIDATING: 'VALIDATING',
+  INVALID: 'INVALID',
+  WARNING: 'WARNING',
+} as const
+export type FieldStateValue = typeof FieldState[keyof typeof FieldState]
+
+export const FieldType = {
+  TEXT: 'TEXT',
+  EMAIL: 'EMAIL',
+  PASSWORD: 'PASSWORD',
+  SELECT: 'SELECT',
+  CHECKBOX: 'CHECKBOX',
+  TEXTAREA: 'TEXTAREA',
+} as const
+export type FieldTypeValue = typeof FieldType[keyof typeof FieldType]
+
+export const FormResultState = {
+  SUCCESS: 'SUCCESS',
+  FAILURE: 'FAILURE',
+} as const
+export type FormResultStateValue = typeof FormResultState[keyof typeof FormResultState]
 
 export interface FormFieldState {
   state: FieldStateValue;
