+++ packages/components/src/Tag.tsx
@@ -2,7 +2,7 @@ import { transparentize } from '@theme-ui/color'
 import Link from 'next/link'
 import { ReactNode } from 'react'
 import { Box, Card, Text } from 'theme-ui'
-import { SxProp } from '../lib/sx'
+import { SxProp } from './sx'
 
 type TagProps={
 	children: ReactNode;
