+++ packages/components/src/viewers.ts
@@ -1,7 +1,8 @@
-import { map } from './map'
-import { markdown } from './markdown'
-import { table } from './table'
-import { text } from './text'
+import map from './map/metadata'
+import markdown from './markdown/metadata'
+import table from './table/metadata'
+import text from './text/metadata'
+
 export const viewers = { markdown, text, map, table } as const
 export const viewerIds = Object.keys(viewers) as ViewerId[]
 export type ViewerId = keyof typeof viewers
